Code - Refactoring
The motivation and goals are:
- better seperate DirectInput from the vehicle dynamics and "foce feedback processing" code - ideally there is only one 10000 in the processing chain at the end of 0:1 or -1:1 scaled value
- factor out parameters to descriptive names
- code changes for readability

+++ src/calculations/lateral_load.h
@@ -9,7 +9,7 @@ void LogMessage(const std::wstring& msg);
 struct CalculatedLateralLoad {
     double steeringDeg = 0.0;
     double speedMph = 0.0;
-    int forceMagnitude = 0;
+    double forceMagnitude = 0;
     int directionVal = 0;
     double lateralG = 0.0;
     double steeringRaw = 0.0;
